fix #6840, bad indexing result type for abstractly-typed index vector
the core problem was that we couldn't infer the type of convert(Int,x::Real),
because there are way too many matching methods. this argues for a
conversion operator that expands to `convert(T,x)::T`, or a way to declare
an overall type for `convert`.

+++ base/operators.jl
@@ -292,7 +292,7 @@ end
 
 # convert to integer index
 to_index(i::Int) = i
-to_index(i::Real) = convert(Int, i)
+to_index(i::Real) = convert(Int,i)::Int
 to_index(r::UnitRange{Int}) = r
 to_index(r::Range{Int}) = r
 to_index(I::UnitRange{Bool}) = find(I)
